[REFACTOR/HM-102]: RoomList 재상용 으로인한 변경
